Change editor path to bundle
