+++ frontend/src/App.js
@@ -8,26 +8,26 @@ import { reduxConstants } from './redux';
 import OperatorHub from './pages/operatorHub/OperatorHub';
 import OperatorPage from './pages/operatorPage/OperatorPage';
 import OperatorPreviewPage from './pages/operatorPreviewPage/OperatorPreviewPage';
-import OperatorFormEditorPage from './pages/OperatorEditorPage/OperatorFormEditorPage';
-import OperatorYamlEditorPage from './pages/OperatorEditorPage/OperatorYamlEditorPage';
+import OperatorBundlePage from './pages/operatorBundlePage/OperatorBundlePage';
+import OperatorYamlEditorPage from './pages/operatorBundlePage/OperatorYamlEditorPage';
 import GettingStarted from './pages/documentation/GettingStarted';
 import WhatIsAnOperator from './pages/documentation/WhatIsAnOperator';
 import Contribute from './pages/documentation/Contribute';
 import HowToInstallOperators from './pages/documentation/HowToInstallOperators';
 import About from './pages/documentation/About';
 import ConfirmationModal from './components/modals/ConfirmationModal';
-import OperatorMetadataPage from './pages/OperatorEditorPage/OperatorMetadataPage';
-import OperatorDeploymentsPage from './pages/OperatorEditorPage/OperatorDeploymentsPage';
-import OperatorInstallModesPage from './pages/OperatorEditorPage/OperatorInstallModesPage';
-import OperatorOwnedCRDsPage from './pages/OperatorEditorPage/OperatorOwnedCRDsPage';
-import OperatorRequiredCRDsPage from './pages/OperatorEditorPage/OperatorRequiredCRDsPage';
-import OperatorPermissionsPage from './pages/OperatorEditorPage/OperatorPermissionsPage';
-import OperatorClusterPermissionsPage from './pages/OperatorEditorPage/OperatorClusterPermissionsPage';
-import OperatorOwnedCRDEditPage from './pages/OperatorEditorPage/OperatorOwnedCRDEditPage';
-import OperatorDeploymentEditPage from './pages/OperatorEditorPage/OperatorDeploymentEditPage';
-import OperatorRequiredCRDEditPage from './pages/OperatorEditorPage/OperatorRequiredCRDEditPage';
-import OperatorPermissionsEditPage from './pages/OperatorEditorPage/OperatorPermissionsEditPage';
-import OperatorClusterPermissionsEditPage from './pages/OperatorEditorPage/OperatorClusterPermissionsEditPage';
+import OperatorMetadataPage from './pages/operatorBundlePage/OperatorMetadataPage';
+import OperatorDeploymentsPage from './pages/operatorBundlePage/OperatorDeploymentsPage';
+import OperatorInstallModesPage from './pages/operatorBundlePage/OperatorInstallModesPage';
+import OperatorOwnedCRDsPage from './pages/operatorBundlePage/OperatorOwnedCRDsPage';
+import OperatorRequiredCRDsPage from './pages/operatorBundlePage/OperatorRequiredCRDsPage';
+import OperatorPermissionsPage from './pages/operatorBundlePage/OperatorPermissionsPage';
+import OperatorClusterPermissionsPage from './pages/operatorBundlePage/OperatorClusterPermissionsPage';
+import OperatorOwnedCRDEditPage from './pages/operatorBundlePage/OperatorOwnedCRDEditPage';
+import OperatorDeploymentEditPage from './pages/operatorBundlePage/OperatorDeploymentEditPage';
+import OperatorRequiredCRDEditPage from './pages/operatorBundlePage/OperatorRequiredCRDEditPage';
+import OperatorPermissionsEditPage from './pages/operatorBundlePage/OperatorPermissionsEditPage';
+import OperatorClusterPermissionsEditPage from './pages/operatorBundlePage/OperatorClusterPermissionsEditPage';
 
 class App extends React.Component {
   constructor(props) {
@@ -46,23 +46,23 @@ class App extends React.Component {
           <Route path="/operator/:channel/:operatorId" component={OperatorPage} />
           <Route path="/operator/:operatorId" component={OperatorPage} />
           <Route path="/preview" component={OperatorPreviewPage} />
-          <Route path="/editor/metadata" component={OperatorMetadataPage} />
-          <Route path="/editor/owned-crds/:crd" component={OperatorOwnedCRDEditPage} />
-          <Route path="/editor/owned-crds" component={OperatorOwnedCRDsPage} />
-          <Route path="/editor/required-crds/:crd" component={OperatorRequiredCRDEditPage} />
-          <Route path="/editor/required-crds" component={OperatorRequiredCRDsPage} />
-          <Route path="/editor/deployments/:deployment" component={OperatorDeploymentEditPage} />
-          <Route path="/editor/deployments" component={OperatorDeploymentsPage} />
-          <Route path="/editor/permissions/:serviceAccountName" component={OperatorPermissionsEditPage} />
-          <Route path="/editor/permissions" component={OperatorPermissionsPage} />
+          <Route path="/bundle/metadata" component={OperatorMetadataPage} />
+          <Route path="/bundle/owned-crds/:crd" component={OperatorOwnedCRDEditPage} />
+          <Route path="/bundle/owned-crds" component={OperatorOwnedCRDsPage} />
+          <Route path="/bundle/required-crds/:crd" component={OperatorRequiredCRDEditPage} />
+          <Route path="/bundle/required-crds" component={OperatorRequiredCRDsPage} />
+          <Route path="/bundle/deployments/:deployment" component={OperatorDeploymentEditPage} />
+          <Route path="/bundle/deployments" component={OperatorDeploymentsPage} />
+          <Route path="/bundle/permissions/:serviceAccountName" component={OperatorPermissionsEditPage} />
+          <Route path="/bundle/permissions" component={OperatorPermissionsPage} />
           <Route
-            path="/editor/cluster-permissions/:serviceAccountName"
+            path="/bundle/cluster-permissions/:serviceAccountName"
             component={OperatorClusterPermissionsEditPage}
           />
-          <Route path="/editor/cluster-permissions" component={OperatorClusterPermissionsPage} />
-          <Route path="/editor/install-modes" component={OperatorInstallModesPage} />
-          <Route path="/editor/yaml" component={OperatorYamlEditorPage} />
-          <Route path="/editor" component={OperatorFormEditorPage} />
+          <Route path="/bundle/cluster-permissions" component={OperatorClusterPermissionsPage} />
+          <Route path="/bundle/install-modes" component={OperatorInstallModesPage} />
+          <Route path="/bundle/yaml" component={OperatorYamlEditorPage} />
+          <Route path="/bundle" component={OperatorBundlePage} />
           <Route path="/getting-started" component={GettingStarted} />
           <Route path="/what-is-an-operator" component={WhatIsAnOperator} />
           <Route path="/contribute" component={Contribute} />
