+++ frontend/src/pages/operatorBundlePage/OperatorClusterPermissionsPage.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import OperatorPermissionsPage from './OperatorPermissionsPage';
-import { sectionsFields } from './editorPageUtils';
+import { sectionsFields } from './bundlePageUtils';
 
 const permissionFields = sectionsFields['cluster-permissions'];
