+++ frontend/src/pages/operatorBundlePage/OperatorRequiredCRDEditPage.js
@@ -3,52 +3,23 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { connect } from 'react-redux';
 import * as _ from 'lodash-es';
-import { safeDump, safeLoad } from 'js-yaml';
 
 import { helpers } from '../../common/helpers';
 import { reduxConstants } from '../../redux';
 import OperatorEditorSubPage from './OperatorEditorSubPage';
-import ResourcesEditor from '../../components/editor/ResourcesEditor';
 import {
   operatorFieldDescriptions,
   operatorFieldPlaceholders,
   operatorFieldValidators
 } from '../../utils/operatorDescriptors';
-import DescriptorsEditor from '../../components/editor/DescriptorsEditor';
-import YamlViewer from '../../components/YamlViewer';
-import { EDITOR_STATUS, getUpdatedFormErrors, sectionsFields } from './editorPageUtils';
+import { EDITOR_STATUS, getUpdatedFormErrors, sectionsFields } from './bundlePageUtils';
 
-const crdsField = sectionsFields['owned-crds'];
+const crdsField = sectionsFields['required-crds'];
 
-const specCapabilities = [
-  'urn:alm:descriptor:com.tectonic.ui:podCount',
-  'urn:alm:descriptor:com.tectonic.ui:endpointList',
-  'urn:alm:descriptor:com.tectonic.ui:label',
-  'urn:alm:descriptor:com.tectonic.ui:resourceRequirements',
-  'urn:alm:descriptor:com.tectonic.ui:selector:',
-  'urn:alm:descriptor:com.tectonic.ui:namespaceSelector',
-  'urn:alm:descriptor:io.kubernetes:',
-  'urn:alm:descriptor:com.tectonic.ui:booleanSwitch'
-];
-
-const statusCapabilities = [
-  'urn:alm:descriptor:com.tectonic.ui:podStatuses',
-  'urn:alm:descriptor:com.tectonic.ui:podCount',
-  'urn:alm:descriptor:org.w3:link',
-  'urn:alm:descriptor:io.kubernetes.conditions',
-  'urn:alm:descriptor:text',
-  'urn:alm:descriptor:prometheusEndpoint',
-  'urn:alm:descriptor:io.kubernetes.phase',
-  'urn:alm:descriptor:io.kubernetes.phase:reason',
-  'urn:alm:descriptor:io.kubernetes:'
-];
-
-class OperatorOwnedCRDEditPage extends React.Component {
+class OperatorRequiredCRDEditPage extends React.Component {
   state = {
     crd: null,
-    crdErrors: null,
-    crdTemplateYaml: '',
-    crdTemplateYamlError: ''
+    crdErrors: null
   };
 
   dirtyFields = {};
@@ -74,30 +45,14 @@ class OperatorOwnedCRDEditPage extends React.Component {
       crd.name = name;
     }
 
-    this.crdIndex = operatorCRDs.indexOf(crd);
-    const kind = _.get(crd, 'kind');
-    const examples = _.get(operator, 'metadata.annotations.alm-examples');
-    const crdTemplates = this.convertExamplesToObj(examples);
-
-    let crdTemplateYaml = '';
-    const crdTemplate = _.find(crdTemplates, { kind });
-    if (crdTemplate) {
-      try {
-        crdTemplateYaml = safeDump(crdTemplate);
-      } catch (e) {
-        console.error(`Unable to convert alm-examples to YAML: ${e}`);
-        crdTemplateYaml = '';
-      }
-    }
-
-    this.section = crdsField === 'owned' ? 'owned-crds' : 'required-crds';
     this.originalName = crd.name;
+    this.crdIndex = operatorCRDs.indexOf(crd);
 
     const errors = getUpdatedFormErrors(operator, formErrors, crdsField, storeEditorFormErrors);
     this.updateCrdErrors(errors);
     storeEditorFormErrors(errors);
 
-    this.setState({ crd, crdTemplateYaml });
+    this.setState({ crd });
 
     if (crd.name === 'new-crd') {
       setTimeout(() => {
@@ -122,25 +77,10 @@ class OperatorOwnedCRDEditPage extends React.Component {
     this.setState({ crdErrors: _.get(crdErrors, 'errors') });
 
     if (crdErrors) {
-      setSectionStatus(EDITOR_STATUS.errors, this.section);
+      setSectionStatus(EDITOR_STATUS.errors, 'required-crds');
     } else {
-      setSectionStatus(EDITOR_STATUS.pending, this.section);
+      setSectionStatus(EDITOR_STATUS.pending, 'required-crds');
     }
-  };
-
-  convertExamplesToObj = examples => {
-    let crdTemplates;
-    if (_.isString(examples)) {
-      try {
-        crdTemplates = JSON.parse(examples);
-      } catch (e) {
-        console.error(`Unable to convert alm-examples: ${e}`);
-        crdTemplates = [];
-      }
-    } else {
-      crdTemplates = examples;
-    }
-    return crdTemplates;
   };
 
   updateCRD = (value, field) => {
@@ -148,34 +88,6 @@ class OperatorOwnedCRDEditPage extends React.Component {
 
     _.set(crd, field, value);
     this.forceUpdate();
-  };
-
-  onTemplateYamlChange = yaml => {
-    const { operator, storeEditorOperator } = this.props;
-    const { crd } = this.state;
-
-    try {
-      const template = safeLoad(yaml);
-      const examples = _.get(operator, 'metadata.annotations.alm-examples');
-      const crdTemplates = this.convertExamplesToObj(examples);
-      const index = _.findIndex(crdTemplates, { kind: _.get(crd, 'kind') });
-      let updatedTemplates;
-      if (index >= 0) {
-        updatedTemplates = [...crdTemplates.slice(0, index), template, ...crdTemplates.slice(index + 1)];
-      } else {
-        updatedTemplates = [...crdTemplates, template];
-      }
-      const updatedExamples = JSON.stringify(updatedTemplates);
-
-      const updatedOperator = _.cloneDeep(operator);
-      _.set(updatedOperator, 'metadata.annotations.alm-examples', updatedExamples);
-
-      storeEditorOperator(updatedOperator);
-
-      this.setState({ crdTemplateYaml: yaml, crdTemplateYamlError: '' });
-    } catch (e) {
-      this.setState({ crdTemplateYaml: yaml, crdTemplateYamlError: e.message });
-    }
   };
 
   validateField = field => {
@@ -275,95 +187,30 @@ class OperatorOwnedCRDEditPage extends React.Component {
     );
   };
 
-  renderCRDFields = () => (
-    <React.Fragment>
-      {this.renderCRDInput('Name', 'name', 'text', this.setNameInputRef)}
-      {this.renderCRDInput('Display Name', 'displayName', 'text')}
-      {this.renderCRDInput('Description', 'description', 'text-area')}
-      {this.renderCRDInput('Kind', 'kind', 'text')}
-      {this.renderCRDInput('Version', 'version', 'text')}
-    </React.Fragment>
-  );
-
-  renderDescriptors = () => {
-    const { crd, crdErrors } = this.state;
-
-    return (
-      <div className="oh_operator-editor__crd-descriptors">
-        <h3>SpecDescriptors, StatusDescriptors, and ActionDescriptors</h3>
-        <p>{_.get(operatorFieldDescriptions, `${crdsField}.descriptors`)}</p>
-        <DescriptorsEditor
-          crd={crd}
-          title="SpecDescriptors"
-          singular="SpecDescriptor"
-          descriptorsErrors={_.get(crdErrors, 'specDescriptors')}
-          description="A reference to fields in the spec block of an object."
-          onUpdate={() => this.validateField('specDescriptors')}
-          descriptorsField="specDescriptors"
-          descriptorOptions={specCapabilities}
-        />
-        <DescriptorsEditor
-          crd={crd}
-          title="StatusDescriptors"
-          singular="StatusDescriptor"
-          descriptorsErrors={_.get(crdErrors, 'statusDescriptors')}
-          description="A reference to fields in the status block of an object."
-          onUpdate={() => this.validateField('statusDescriptors')}
-          descriptorsField="statusDescriptors"
-          descriptorOptions={statusCapabilities}
-        />
-      </div>
-    );
-  };
-
-  renderTemplates = () => {
-    const { crdTemplateYaml, crdTemplateYamlError } = this.state;
-
-    return (
-      <React.Fragment>
-        <h3>CRD Templates</h3>
-        <p>{_.get(operatorFieldDescriptions, 'metadata.annotations.alm-examples')}</p>
-        <YamlViewer
-          onBlur={yaml => this.onTemplateYamlChange(yaml)}
-          editable
-          yaml={crdTemplateYaml}
-          error={crdTemplateYamlError}
-          allowClear
-        />
-      </React.Fragment>
-    );
-  };
-
   render() {
     const { history } = this.props;
-    const { crd, crdErrors } = this.state;
 
     return (
       <OperatorEditorSubPage
-        title="Edit Owned CRD"
+        title="Edit Required CRD"
         tertiary
-        lastPage="owned-crds"
-        lastPageTitle="Owned CRDs"
+        lastPage="required-crds"
+        lastPageTitle="Required CRDs"
         history={history}
       >
         <form className="oh-operator-editor-form">
-          {this.renderCRDFields()}
-          <ResourcesEditor
-            crd={crd}
-            crdErrors={crdErrors}
-            onUpdate={() => this.validateField('resources')}
-            title="Resources"
-            field={`${crdsField}.resources`}
-          />
-          {this.renderDescriptors()}
-          {this.renderTemplates()}
+          {this.renderCRDInput('Name', 'name', 'text', this.setNameInputRef)}
+          {this.renderCRDInput('Display Name', 'displayName', 'text')}
+          {this.renderCRDInput('Description', 'description', 'text-area')}
+          {this.renderCRDInput('Kind', 'kind', 'text')}
+          {this.renderCRDInput('Version', 'version', 'text')}
         </form>
       </OperatorEditorSubPage>
     );
   }
 }
 
-OperatorOwnedCRDEditPage.propTypes = {
+OperatorRequiredCRDEditPage.propTypes = {
   operator: PropTypes.object,
   formErrors: PropTypes.object,
   sectionStatus: PropTypes.object,
@@ -376,7 +223,7 @@ OperatorOwnedCRDEditPage.propTypes = {
   match: PropTypes.object.isRequired
 };
 
-OperatorOwnedCRDEditPage.defaultProps = {
+OperatorRequiredCRDEditPage.defaultProps = {
   operator: {},
   formErrors: {},
   sectionStatus: {},
@@ -413,4 +260,4 @@ const mapStateToProps = state => ({
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(OperatorOwnedCRDEditPage);
+)(OperatorRequiredCRDEditPage);
