+++ frontend/src/pages/operatorBundlePage/bundlePageUtils.js
