+++ server/routes/uiRoutes.js
@@ -43,7 +43,7 @@ module.exports = app => {
   const uiRoutes = [
     'operator',
     'preview',
-    'editor',
+    'bundle',
     'about',
     'contribute',
     'getting-started',
